Move eve/product_type.hpp to eve/traits
