+++ examples/oop/complex_numbers__declaring_an_object_type.cpp
@@ -25,7 +25,7 @@
 #include <eve/eve.hpp>
 #include <eve/module/core.hpp>
 #include <eve/module/math.hpp>
-#include <eve/product_type.hpp>
+#include <eve/traits/product_type.hpp>
 
 #include <eve/algo/container/soa_vector.hpp>
 #include <eve/algo/reduce.hpp>
