+++ examples/start_here.cpp
@@ -33,7 +33,7 @@ int const* find_negative_number(int const* f, int const* l)
 
 #include <eve/module/core.hpp>
 #include <eve/module/math.hpp>
-#include <eve/product_type.hpp>
+#include <eve/traits/product_type.hpp>
 
 #include <vector>
