+++ include/eve/detail/function/simd/common/deinterleave_groups.hpp
@@ -7,7 +7,7 @@
 //==================================================================================================
 #pragma once
 
-#include <eve/product_type.hpp>
+#include <eve/traits/product_type.hpp>
 #include <eve/module/core/regular/deinterleave_groups_shuffle.hpp>
 #include <eve/module/core/regular/swap_adjacent_groups.hpp>
 #include <eve/module/core/regular/shuffle.hpp>
