+++ include/eve/detail/function/simd/x86/interleave.hpp
@@ -9,7 +9,7 @@
 
 #include <eve/detail/abi.hpp>
 #include <eve/detail/function/bit_cast.hpp>
-#include <eve/product_type.hpp>
+#include <eve/traits/product_type.hpp>
 
 namespace eve::detail
 {
